Renamed packages to singular word
