+++ backend/src/main/java/com/github/redstylzz/backend/repository/IMongoUserRepository.java
@@ -1,6 +1,6 @@
-package com.github.redstylzz.backend.repositories;
+package com.github.redstylzz.backend.repository;
 
-import com.github.redstylzz.backend.models.MongoUser;
+import com.github.redstylzz.backend.model.MongoUser;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Repository;
