+++ backend/src/main/java/com/github/redstylzz/backend/service/LoginService.java
@@ -1,6 +1,6 @@
-package com.github.redstylzz.backend.services;
+package com.github.redstylzz.backend.service;
 
-import com.github.redstylzz.backend.models.MongoUser;
+import com.github.redstylzz.backend.model.MongoUser;
 import org.apache.juli.logging.Log;
 import org.apache.juli.logging.LogFactory;
 import org.springframework.http.HttpStatus;
